Remove ruby and package-cloud from dod-package-dev
The package-cloud gem package does not install on versions of ruby <2.0
so trusty default ruby is not new enough. This results in failures in
auto build:

https://hub.docker.com/r/heroku/dod-package-dev/builds/

This commit removes ruby and the package-cloud gem from the docker
image, as it was included as a convenience. Users can continue to use
the web or host based package-cloud.