Add focal package builder, update PG apt key (#7)
As part of upgrading to focal and bionic, we need a matching builder
image. This was using a hardcoded GPG key(!) that has now expired, so
I've updated the script to pull the latest from the right places. This
causes a divergence from the origin
(`postgresql-common/apt.postgresql.org.sh`), but for our purposes should
be fine.

Ref: https://heroku.slack.com/archives/CH6PAB8HX/p1591890951278500