Add better locale generation
Due to some "interesting" quirks on how multiple entries are created by
`locarchive`, which only triggers when an explicit path is _not_ passed
in, we were generating locales that Postgres couldn't find based on the
information stored in `pg_collation`.

To sidestep this, we patch `localedef` and `locarchive` to point
directly to `/usr/lib/locale` (instead of the `glibc`-local directory)
and patch `locale-gen` to use the patched `localedef`.

Ref: GUS-W-9208215